+++ lib/react_on_rails/test_helper/ensure_assets_compiled.rb
@@ -50,7 +50,7 @@ Detected the following stale generated files:
   #{stale_files.join("\n  ")}
 
 React on Rails will ensure your JavaScript generated files are up to date, using your
-/client level package.json `#{ReactOnRails.configuration.npm_build_test_command}` command.
+`#{ReactOnRails::Utils.prepend_cd_node_modules_directory(ReactOnRails.configuration.build_test_command)}` command.
 
         MSG
         # rubocop:enable Layout/IndentHeredoc