+++ spec/react_on_rails/utils_spec.rb
@@ -1,9 +1,65 @@
 # frozen_string_literal: true
 
 require_relative "spec_helper"
+require "webpacker"
 
 module ReactOnRails
   RSpec.describe Utils do
+    describe ".bundle_js_file_path" do
+      before do
+        allow(ReactOnRails).to receive_message_chain(:configuration, :generated_assets_dir)
+          .and_return("public/webpack/development")
+      end
+
+      subject do
+        Utils.bundle_js_file_path("webpack-bundle.js")
+      end
+
+      context "With Webpacker enabled and file in manifest" do
+        before do
+          allow(Rails).to receive(:root).and_return(Pathname.new("."))
+          allow(Webpacker).to receive_message_chain("dev_server.running?")
+            .and_return(false)
+          allow(Webpacker).to receive_message_chain("config.public_output_path")
+            .and_return("/webpack/development")
+          allow(Webpacker).to receive_message_chain("manifest.lookup")
+            .with("webpack-bundle.js")
+            .and_return("/webpack/development/webpack-bundle-0123456789abcdef.js")
+          allow(Utils).to receive(:using_webpacker?).and_return(true)
+        end
+
+        it { expect(subject).to eq("public/webpack/development/webpack-bundle-0123456789abcdef.js") }
+      end
+
+      context "Without Webpacker enabled" do
+        before { allow(Utils).to receive(:using_webpacker?).and_return(false) }
+
+        it { expect(subject).to eq("public/webpack/development/webpack-bundle.js") }
+      end
+    end
+
+    describe ".server_bundle_js_file_path" do
+      subject do
+        Utils.server_bundle_js_file_path
+      end
+
+      context "With Webpacker enabled and server file not in manifest" do
+        before do
+          allow(Rails).to receive(:root).and_return(Pathname.new("."))
+          allow(ReactOnRails).to receive_message_chain("configuration.server_bundle_js_file")
+            .and_return("webpack-bundle.js")
+          allow(Webpacker).to receive_message_chain("config.public_output_path")
+            .and_return("public/webpack/development")
+          allow(Webpacker).to receive_message_chain("manifest.lookup")
+            .with("webpack-bundle.js")
+            .and_raise(Webpacker::Manifest::MissingEntryError)
+          allow(Utils).to receive(:using_webpacker?).and_return(true)
+        end
+
+        it { expect(subject).to eq("public/webpack/development/webpack-bundle.js") }
+      end
+    end
+
     describe ".wrap_message" do
       subject do
         <<-MSG.strip_heredoc